Move the test into the integration suite.
This lets us use DBPath to clean things up slightly.
