+++ tests/test_slice_transform.rs
@@ -64,3 +64,32 @@ pub fn test_slice_transform() {
         }
     }
 }
+
+#[test]
+fn test_no_in_domain() {
+    fn extract_suffix(slice: &[u8]) -> &[u8] {
+        if slice.len() > 4 {
+            &slice[slice.len() - 4..slice.len()]
+        } else {
+            slice
+        }
+    }
+
+    let db_path = DBPath::new("_rust_rocksdb_prefix_test");
+    {
+        let mut opts = Options::default();
+        opts.create_if_missing(true);
+        opts.set_prefix_extractor(SliceTransform::create(
+            "test slice transform",
+            extract_suffix,
+            None,
+        ));
+        opts.set_memtable_prefix_bloom_ratio(0.1);
+
+        let db = DB::open(&opts, &db_path).unwrap();
+        db.put(b"key_sfx1", b"a").unwrap();
+        db.put(b"key_sfx2", b"b").unwrap();
+
+        assert_eq!(db.get(b"key_sfx1").unwrap().unwrap(), b"a");
+    }
+}
